fix: 修复存储、事务与查询缺陷，新增备份、校验、提交通知等功能
落盘格式不变。

修复：
- 事务：安全点后提交漏打确认标记导致改动丢失；提交失败后头页未还原、实例作废后仍写日志；
  集合表写满不再作废实例；诊断视图并发读 map 崩溃；Close 作废在途事务与迭代器；
  检查点被持续读事务饿死；页缓存改为分片 CLOCK
- 引擎：Update/Upsert 失败时文档与索引保持原状；自增序列被手写主键压低及溢出回绕；
  同一集合多条向量索引插入报错；向量索引维护改为批量定位
- 查询：分组查询的 ORDER BY 被索引误吃；UpdateMany/DeleteMany 分批改走主键区间；
  Include 引用展开加查询内缓存；外部排序内存估算
- 文档与表达式：解码与编码嵌套深度上限；数值比较、字段名查找、正则、LIKE、方法调用、
  忽略大小写比较去掉热路径分配
- 公开层：$file 默认限制在工作目录、$query 只放行 SELECT；共享模式同句柄并发与锁名；
  重建原子替换与失败后恢复句柄；文件存储写入、删除、改元信息互斥，清理孤块

新增：
- DB.Backup / DB.BackupFile 在线一致性备份
- DB.Stats 与 WithLogger
- DB.OnCommit 提交后变更通知
- Collection.EnsureTTL 过期文档自动删除
- Verify / DB.Verify 只读完整性校验（含多键索引条数）

+++ internal/xquery/index.go
@@ -612,7 +612,7 @@ func (o *opIn) keyOrdered() bool { return false }
 
 // cost 是单次等值查的代价乘以值的个数。
 func (o *opIn) cost(ix *xpage.CollectionIndex) uint32 {
-	unit := uint32(costEquals)
+	unit := costEquals
 	if ix.Unique {
 		unit = costEqualsUnique
 	}
